backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
cryptography==46.0.1
deprecation==2.1.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==2.20.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
realtime==2.20.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
websockets==15.0.1
